+++ cost-payment-service/src/main/java/com/example/costpayment/dto/FundSummaryDto.java
@@ -3,6 +3,7 @@ package com.example.costpayment.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import java.math.BigDecimal;
 
 /**
  * DTO: Tổng quan quỹ chung
@@ -14,16 +15,16 @@ public class FundSummaryDto {
 
     private Integer fundId;
     private Integer groupId;
-    private Double totalContributed; // Tổng đóng góp
-    private Double currentBalance; // Số dư hiện tại
-    private Double totalDeposit; // Tổng nạp
-    private Double totalWithdraw; // Tổng rút
+    private BigDecimal totalContributed; // Tổng đóng góp
+    private BigDecimal currentBalance; // Số dư hiện tại
+    private BigDecimal totalDeposit; // Tổng nạp
+    private BigDecimal totalWithdraw; // Tổng rút
     private Long pendingRequests; // Số yêu cầu chờ duyệt
     private String updatedAt;
 
     public FundSummaryDto(
         Integer fundId, Integer groupId, 
-        Double totalContributed, Double currentBalance,
+        BigDecimal totalContributed, BigDecimal currentBalance,
         Long pendingRequests, String updatedAt
     ) {
         this.fundId = fundId;
